requirements_old.txt
# Anaconda Python 3.7.4 (default, Aug 13 2019, 20:35:49)
python-chess==0.28.3
shell==1.0.1
joblib==0.14.0
numpy==1.17.4
pandas==0.25.3
scikit-learn==0.21.3
tensorflow-gpu==2.1.0
tensorboard==2.1.0
tensorflow-estimator==2.1.0
grpcio==1.27.2
termcolor==1.1.0
psutil==5.7.0
dispy==4.11.1
fire==0.2.1
docopt==0.6.2
# minimalcluster
